Make RunREDUCEPrefs.save update the appropriate RunREDUCEPrefs field when it is passed a (second) value argument. Remove all saved versions before saving the current versions to avoid leaving obsolete data saved. Do not explicitly delete any redundant saved command args since this is no longer relevant.
Introduce methods to build the REDUCE version submenus, which can be called also when the REDUCEConfigDialog is saved, in order to update the submenus in case the REDUCE version list has changed in some way. Correct the action when a JRadioButtonMenuItem is selected so that it saves the appropriate version name.